server/requirements.txt
fastapi
pydantic
python-dotenv
openai
uvicorn
torch
transformers
sentence-transformers
numpy